Remove repr, cleanup dependency, combine logs
Per @annie-xd-wang's requests...
