+++ src/navigate/controller/sub_controllers/stages.py
@@ -39,11 +39,12 @@ import logging
 
 # Local Imports
 from navigate.controller.sub_controllers.gui import GUIController
-from navigate.model.devices import log_initialization
+from navigate.tools.decorators import log_initialization
 
 # Logger Setup
 p = __name__.split(".")[1]
 logger = logging.getLogger(p)
+
 
 @log_initialization
 class StageController(GUIController):
