+++ src/navigate/log_files/logging.yml
@@ -9,8 +9,6 @@ filters:
     (): ext://navigate.log_files.filters.PerformanceFilter
   not_performance:
     (): ext://navigate.log_files.filters.NonPerfFilter
-  only_debug:
-    (): ext://navigate.log_files.filters.DebugFilter
 handlers:
   console:
     class: logging.StreamHandler
@@ -30,29 +28,19 @@ handlers:
     level: DEBUG
     formatter: base
     filename: view_controller_debug.log
-    filters: [not_performance, only_debug]
+    filters: [not_performance]
     mode: a
-#  performance:
-#    class: logging.handlers.RotatingFileHandler
-#    level: DEBUG
-#    formatter: base
-#    filename: view_controller_performance.log
-#    filters: [performance_specs]
-#    mode: a
 loggers:
   view:
     level: DEBUG
-#    handlers: [console, vc_info, vc_debug, performance]
     handlers: [console, vc_info, vc_debug]
     propagate: no
   controller:
     level: DEBUG
-#    handlers: [console, vc_info, vc_debug, performance]
     handlers: [console, vc_info, vc_debug]
     propagate: no
   config:
     level: DEBUG
-#    handlers: [console, vc_info, vc_debug, performance]
     handlers: [console, vc_info, vc_debug]
     propagate: no
