+++ src/navigate/log_files/model_logging.yml
@@ -9,8 +9,6 @@ filters:
     (): ext://navigate.log_files.filters.PerformanceFilter
   not_performance:
     (): ext://navigate.log_files.filters.NonPerfFilter
-  only_debug:
-    (): ext://navigate.log_files.filters.DebugFilter
 handlers:
   console:
     class: logging.StreamHandler
@@ -30,7 +28,7 @@ handlers:
     level: DEBUG
     formatter: base
     filename: model_debug.log
-    filters: [not_performance, only_debug]
+    filters: [not_performance]
     mode: a
   model_error:
     class: logging.handlers.RotatingFileHandler
@@ -39,13 +37,6 @@ handlers:
     filename: model_error.log
     filters: [not_performance]
     mode: a
-#  model_performance:
-#    class: logging.handlers.RotatingFileHandler
-#    level: DEBUG
-#    formatter: base
-#    filename: model_performance.log
-#    filters: [performance_specs]
-#    mode: a
 loggers:
   model:
     level: DEBUG
