+++ src/navigate/model/devices/camera/hamamatsu.py
@@ -37,7 +37,8 @@ import logging
 
 # Local Imports
 from navigate.model.devices.camera.base import CameraBase
-from navigate.model.devices import log_initialization
+from navigate.tools.decorators import log_initialization
+
 # Logger Setup
 p = __name__.split(".")[1]
 logger = logging.getLogger(p)
