+++ src/navigate/model/devices/camera/synthetic.py
@@ -42,7 +42,7 @@ from tifffile import TiffFile, TiffFileError
 # Local Imports
 from navigate.model.analysis import camera
 from navigate.model.devices.camera.base import CameraBase
-from navigate.model.devices import log_initialization
+from navigate.tools.decorators import log_initialization
 
 # Logger Setup
 p = __name__.split(".")[1]
