+++ src/navigate/model/devices/daq/base.py
@@ -37,11 +37,12 @@ import logging
 
 # Local Imports
 from navigate.model.waveforms import camera_exposure
-from navigate.model.devices import log_initialization
+from navigate.tools.decorators import log_initialization
 
 # Logger Setup
 p = __name__.split(".")[1]
 logger = logging.getLogger(p)
+
 
 @log_initialization
 class DAQBase:
