+++ src/navigate/model/devices/filter_wheel/base.py
@@ -36,7 +36,7 @@ import logging
 # Third Party Imports
 
 # Local Imports
-from navigate.model.devices import log_initialization
+from navigate.tools.decorators import log_initialization
 
 # Logger Setup
 p = __name__.split(".")[1]
