+++ src/navigate/model/devices/filter_wheel/ni.py
@@ -41,7 +41,7 @@ from nidaqmx.constants import LineGrouping
 
 # Local Imports
 from navigate.model.devices.filter_wheel.base import FilterWheelBase
-from navigate.model.devices import log_initialization
+from navigate.tools.decorators import log_initialization
 
 # Logger Setup
 p = __name__.split(".")[1]
