+++ src/navigate/model/devices/galvo/base.py
@@ -37,7 +37,7 @@ import logging
 
 # Local Imports
 from navigate.model.waveforms import sawtooth, sine_wave
-from navigate.model.devices import log_initialization
+from navigate.tools.decorators import log_initialization
 
 # # Logger Setup
 p = __name__.split(".")[1]
