+++ src/navigate/model/devices/mirrors/imop.py
@@ -38,7 +38,7 @@ import logging
 
 # Local Imports
 from navigate.model.devices.mirrors.base import MirrorBase
-from navigate.model.devices import log_initialization
+from navigate.tools.decorators import log_initialization
 
 # Logger Setup
 p = __name__.split(".")[1]
