+++ src/navigate/model/devices/remote_focus/base.py
@@ -41,8 +41,7 @@ from navigate.model.waveforms import (
     smooth_waveform,
     remote_focus_ramp_triangular,
 )
-from navigate.model.devices import log_initialization
-
+from navigate.tools.decorators import log_initialization
 
 # # Logger Setup
 p = __name__.split(".")[1]
