+++ src/navigate/model/devices/remote_focus/synthetic.py
@@ -37,7 +37,7 @@ import logging
 
 # Local Imports
 from navigate.model.devices.remote_focus.base import RemoteFocusBase
-from navigate.model.devices import log_initialization
+from navigate.tools.decorators import log_initialization
 
 # # Logger Setup
 p = __name__.split(".")[1]
