+++ src/navigate/model/devices/shutter/synthetic.py
@@ -38,7 +38,7 @@ import logging
 
 # Local Imports
 from navigate.model.devices.shutter.base import ShutterBase
-from navigate.model.devices import log_initialization
+from navigate.tools.decorators import log_initialization
 
 # Logger Setup
 p = __name__.split(".")[1]
