+++ src/navigate/model/devices/stages/asi.py
@@ -41,8 +41,7 @@ from navigate.model.devices.APIs.asi.asi_tiger_controller import (
     TigerController,
     TigerException,
 )
-from navigate.model.devices import log_initialization
-
+from navigate.tools.decorators import log_initialization
 
 # Logger Setup
 p = __name__.split(".")[1]
