+++ src/navigate/model/devices/stages/asi_MSTwoThousand.py
@@ -41,7 +41,7 @@ from navigate.model.devices.APIs.asi.asi_MS2000_controller import (
     MS2000Controller,
     MS2000Exception,
 )
-from navigate.model.devices import log_initialization
+from navigate.tools.decorators import log_initialization
 
 # Logger Setup
 p = __name__.split(".")[1]
