+++ src/navigate/model/devices/stages/base.py
@@ -36,7 +36,7 @@ from multiprocessing.managers import ListProxy
 # Third Party Imports
 
 # Local Imports
-from navigate.model.devices import log_initialization
+from navigate.tools.decorators import log_initialization
 
 # Logger Setup
 p = __name__.split(".")[1]
