+++ src/navigate/model/devices/stages/ni.py
@@ -41,7 +41,7 @@ import nidaqmx
 
 # Local Imports
 from navigate.model.devices.stages.base import StageBase
-from navigate.model.devices import log_initialization
+from navigate.tools.decorators import log_initialization
 
 # Logger Setup
 p = __name__.split(".")[1]
