+++ src/navigate/model/devices/stages/synthetic.py
@@ -37,7 +37,7 @@ import time
 
 # Local Imports
 from navigate.model.devices.stages.base import StageBase
-from navigate.model.devices import log_initialization
+from navigate.tools.decorators import log_initialization
 
 # Logger Setup
 p = __name__.split(".")[1]
