+++ src/navigate/model/devices/stages/tl_kcube_inertial.py
@@ -37,8 +37,7 @@ from multiprocessing.managers import ListProxy
 
 # Local Imports
 from navigate.model.devices.stages.base import StageBase
-from navigate.model.devices import log_initialization
-
+from navigate.tools.decorators import log_initialization
 
 # Logger Setup
 p = __name__.split(".")[1]
