+++ src/navigate/model/devices/zoom/base.py
@@ -36,11 +36,12 @@ import logging
 # Third Party Imports
 
 # Local Imports
-from navigate.model.devices import log_initialization
+from navigate.tools.decorators import log_initialization
 
 # Logger Setup
 p = __name__.split(".")[1]
 logger = logging.getLogger(p)
+
 
 @log_initialization
 class ZoomBase:
