+++ src/navigate/model/devices/zoom/dynamixel.py
@@ -38,7 +38,7 @@ import time
 # Local Imports
 from navigate.model.devices.APIs.dynamixel import dynamixel_functions as dynamixel
 from navigate.model.devices.zoom.base import ZoomBase
-from navigate.model.devices import log_initialization
+from navigate.tools.decorators import log_initialization
 
 # Logger Setup
 p = __name__.split(".")[1]
