+++ src/navigate/model/devices/zoom/synthetic.py
@@ -38,11 +38,12 @@ import logging
 
 # Local Imports
 from navigate.model.devices.zoom.base import ZoomBase
-from navigate.model.devices import log_initialization
+from navigate.tools.decorators import log_initialization
 
 # Logger Setup
 p = __name__.split(".")[1]
 logger = logging.getLogger(p)
+
 
 @log_initialization
 class SyntheticZoom(ZoomBase):
